Splitting up source files to smaller ones.
Splitting up source files into smaller more manageable ones.
Also changed API for RTOS_CreateTask() to take task name as well.
Separate RTOS_KillSelf() and RTOS_SuspendSelf() functions from
RTOS_KillTask() and RTOS_Suspend().
A number of configuration options have also changed, need to update
documentation on those.
