+++ examples/timeshare/rtos_config.h
@@ -36,7 +36,7 @@ extern "C" {
 #define RTOS_INCLUDE_WAKEUP 
 #define RTOS_INCLUDE_CHANGEPRIORITY
 
-// #define RTOS_USE_TIMER_TASK
+#define RTOS_USE_TIMER_TASK
 
 #define RTOS_TASK_NAME_LENGTH		32
