+++ rtos/rtos_internals.h
@@ -1,7 +1,7 @@
 #ifndef RTOS_INTERNALS_H
 #define RTOS_INTERNALS_H
 /*
-* Copyright (c) Andras Zsoter 2014-2015.
+* Copyright (c) Andras Zsoter 2014-2016.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -38,6 +38,8 @@ extern "C" {
 #define RTOS_TaskSet_Difference(S1, S2) ((S1) & (~(S2))) /* Set-theoric difference AKA Relative Complement. */
 #endif
 
+#define rtos_TaskFromPriority(PRIORITY) (((PRIORITY) > RTOS_Priority_Highest) ? (RTOS_Task *)0 : RTOS.TaskList[priority])
+
 #if defined(RTOS_INCLUDE_SCHEDULER_LOCK)
 #define RTOS_SchedulerIsLocked() (0 != RTOS.SchedulerLocked)
 #else
@@ -48,13 +50,34 @@ extern "C" {
 extern void rtos_TargetInitializeTask(RTOS_Task *task, unsigned long stackCapacity);
 
 // Initialization.
-extern RTOS_RegInt rtos_CreateTask(RTOS_Task *task, RTOS_TaskPriority priority, void *sp0, unsigned long stackCapacity, void (*f)(), void *param);
+extern RTOS_RegInt rtos_CreateTask(RTOS_Task *task, void *sp0, unsigned long stackCapacity, void (*f)(), void *param);
+extern RTOS_RegInt rtos_RegisterTask(RTOS_Task *task, RTOS_TaskPriority priority);
 
+// Internal task states.
+#define RTOS_TASK_STATUS_ACTIVE		 ((RTOS_RegInt)0)
+#define RTOS_TASK_STATUS_WAITING	 ((RTOS_RegInt)1)
+#define RTOS_TASK_STATUS_SLEEPING	 ((RTOS_RegInt)2)
+#define RTOS_TASK_STATUS_TIMED_OUT 	 ((RTOS_RegInt)3)
+#define RTOS_TASK_STATUS_AWAKENED	 ((RTOS_RegInt)4)
+#define RTOS_TASK_STATUS_PREEMPTED_FLAG	 ((RTOS_RegInt)0x40)
+#define RTOS_TASK_STATUS_SUSPENDED_FLAG	 ((RTOS_RegInt)0x80)
+// #define RTOS_TASK_STATUS_STOPPED	 ((RTOS_RegInt)0x07)	  
+#define RTOS_TASK_STATUS_KILLED 	 ((RTOS_RegInt)0x0F)
 
+#if defined(RTOS_SUPPORT_EVENTS)
+extern RTOS_RegInt rtos_MapStatusToReturnValue(RTOS_RegInt status);
+#endif
+
+#if defined(RTOS_INCLUDE_WAKEUP)
+extern RTOS_RegInt rtos_WakeupTask(RTOS_Task *task);
+#endif
 
 #if defined(RTOS_SMP)
 #define RTOS_CURRENT_TASK() RTOS.CurrentTasks[RTOS_CurrentCpu()]
 #define rtos_IsCpuInsideIsr(CPU) (0 != RTOS.InterruptNesting[(CPU)])
+// extern RTOS_RegInt rtos_TaskForceInterrupted(RTOS_Task *task);
+extern void rtos_RestrictPriorityToCpus(RTOS_TaskPriority priority, RTOS_CpuMask cpus);
+
 #else
 #define RTOS_CURRENT_TASK() RTOS.CurrentTask
 #endif
@@ -65,11 +88,15 @@ extern void rtos_Scheduler(void);
 extern void rtos_SchedulerForYield(void);
 extern void rtos_RunTask(void);
 
+// Only to be called by OS components.
+extern RTOS_RegInt rtos_SignalEvent(RTOS_EventHandle *event);
+extern void rtos_WaitForEvent(RTOS_EventHandle *event, RTOS_Task *task, RTOS_Time timeout);
+
 #if defined(RTOS_SUPPORT_TIMESHARE)
-// Doubly-Linked Lists.
-extern void rtos_AppendToTaskDList(volatile RTOS_Task_DLList *list, RTOS_RegInt which, RTOS_Task *task);
-extern void rtos_RemoveFromTaskDList(volatile RTOS_Task_DLList *list, RTOS_RegInt which, RTOS_Task *task);
-extern RTOS_Task *rtos_RemoveFirstTaskFromDList(volatile RTOS_Task_DLList *list, RTOS_RegInt which);
+// Doubly Linked Lists.
+extern void rtos_AppendToTaskDLList(volatile RTOS_Task_DLList *list, RTOS_RegInt which, RTOS_Task *task);
+extern void rtos_RemoveFromTaskDLList(volatile RTOS_Task_DLList *list, RTOS_RegInt which, RTOS_Task *task);
+extern RTOS_Task *rtos_RemoveFirstTaskFromDLList(volatile RTOS_Task_DLList *list, RTOS_RegInt which);
 
 extern void rtos_PreemptTask(RTOS_Task *task);
 extern void rtos_SchedulePeer(void);
